[#26] Domain::Shared::Entity extracted from duplicities
